+++ src/app/components/cotizaciones/chart-line/chart-cotizaciones.component.css
