+++ src/app/components/cotizaciones/chart-line/chart-cotizaciones.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ChartCotizacionesComponent } from './chart-cotizaciones.component';
+import { ChartCotizacionesComponent } from '../chart-cotizaciones.component';
 
 describe('ChartCotizacionesComponent', () => {
   let component: ChartCotizacionesComponent;
