--- src/app/models/services/handle-date-time-value.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { HandleDateTimeValueService } from './handle-date-time-value.service';
-
-describe('HandleDateTimeValueService', () => {
-  let service: HandleDateTimeValueService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(HandleDateTimeValueService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
