+++ src/app/utils/util.spec.ts
@@ -1,0 +1,7 @@
+import { Util } from './util';
+
+describe('HandleDateTimeValue', () => {
+  it('should create an instance', () => {
+    expect(new Util()).toBeTruthy();
+  });
+});
